Tweak append arg
